pyproject.toml
[tool.poetry]
name = "webhooks"
version = "0.1.0"
description = ""
authors = ["Johnny Rocketfingers <developer94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
celery = "^4.3"
django-rest-framework = "^0.1.0"
redis = "^3.3"
requests = "^2.22"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-django = "^3.5"
pytest-cov = "^2.7"
flake8 = "^3.7"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
isort = "^4.3"
model_mommy = "^1.6"
ipython = "^7.7"
requests-mock = "^1.6"
pdbpp = "^0.10.0"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
